refactor: Remove getColumns() method from the schema in favor of com.jamhour.data package objects; add auto_increment to id column in the course table
This commit removes the unwanted methods from the schema class that were redundant with the methods in the com.jamhour.data package objects. This promotes the use of the data objects as the main interface for the database operations. It also adds the auto_increment attribute to the id column in the course table to simplify the insertion of new courses.